Move posts into posts dir and fix a cache issue
